BREAKING CHANGE
Changed variable type from a list to a string for architecture. Even
though the Terraform variable takes a list it has to be a single item in
the list. So we make it a string and convert it to a single item list.

Validation was added to enforce constraints on the correct
architectures.