add: Ability to Attach Multiple Resource Policies
You can now add mutliple resource polcicies using the new variable
"resource_policies" which takes a map of objects with values for
action, principal, source_arn, the key will be used for the
statement id. So please keep that in mind and do not go crazy
with names and their length.